OCPRHV-619: Installcfg, Add AddProviderPlatformToInstallConfig
The AddProviderPlatformToInstallConfig is used to add a platform managed by a registered provider to an Install Config struct.
 That way each provider can implement its own specific parts.
 We also move existing functions/structs from installcfg.go to each provider.
 Only changes made from the original function is instead of recalculating the number of host we take the value from the cluster struct.
Also, we separate the installcfg builder into its own package - this is a general best practice and was needed so the implementation of the installcfg could import the provider registry and avoid a dependency cycle.
* dependency cycle: installcfg.go imported the providers (baremetal, vsphere and registry) and they imported it back, go doesn't allow it.
